+++ src/scripts/components/input-controls/textfield.vue
@@ -131,6 +131,10 @@ export default {
   },
   props: {
     // UI variants
+    type: {
+      type: [String, Number],
+      default: 0
+    },
     outlined: {
       type: Boolean,
       default: false