+++ src/scripts/components/selection-controls/select.vue
@@ -11,6 +11,7 @@
       </slot>
       <i class="mdc-select__dropdown-icon"></i>
       <div
+        ref="text"
         class="mdc-select__selected-text"
         role="button"
         aria-haspopup="listbox"
@@ -20,19 +21,19 @@
         :aria-controls="helperTextId"
         :aria-describedby="helperTextId"
       >
-        {{ selectedIndex === -1 ? '' : currentOption[optionLabel] }}
+        {{ currentSelectedLabel }}
       </div>
       <div v-if="isOutlined" class="mdc-notched-outline">
         <div class="mdc-notched-outline__leading"></div>
         <div class="mdc-notched-outline__notch">
-          <ui-floating-label v-if="!noLabel">
+          <ui-floating-label v-if="!noLabel" :isFloatAbove="isFloatAbove">
             <slot>{{ label }}</slot>
           </ui-floating-label>
         </div>
         <div class="mdc-notched-outline__trailing"></div>
       </div>
       <template v-else>
-        <ui-floating-label v-if="!noLabel">
+        <ui-floating-label v-if="!noLabel" :isFloatAbove="isFloatAbove">
           <slot>{{ label }}</slot>
         </ui-floating-label>
         <div class="mdc-line-ripple"></div>
@@ -42,29 +43,18 @@
     <div class="mdc-select__menu mdc-menu mdc-menu-surface" role="listbox">
       <ul class="mdc-list">
         <li
-          v-if="defaultLabel"
-          :class="[
-            'mdc-list-item',
-            { 'mdc-list-item--selected': defaultValue === selectedValue }
-          ]"
-          :data-value="defaultValue"
-          aria-selected="true"
-          role="option"
-        >
-          {{ defaultLabel }}
-        </li>
-        <li
-          v-for="(option, index) in options"
+          v-for="(option, index) in currentOptions"
           :key="index"
           :class="[
             'mdc-list-item',
             {
-              'mdc-list-item--selected': option[optionValue] === selectedValue,
+              'mdc-list-item--selected':
+                option[optionValue] === currentSelectedValue,
               'mdc-list-item--disabled': option.disabled
             }
           ]"
           :data-value="option[optionValue]"
-          :aria-selected="option[optionValue] === selectedValue"
+          :aria-selected="option[optionValue] === currentSelectedValue"
           :aria-disabled="option.disabled"
           role="option"
         >
@@ -92,9 +82,11 @@ const UI_SELECT = {
     icon: 'mdc-select__icon'
   },
   EVENT: {
+    UPDATE_MODEL: 'update-value',
     CHANGE: 'change',
     SELECTED: 'selected'
-  }
+  },
+  DEFAULT_SELECTED_INDEX: -1
 };
 
 export default {
@@ -105,20 +97,20 @@ export default {
   mixins: [typeMixin, materialIconMixin],
   model: {
     prop: 'model',
-    event: UI_SELECT.EVENT.CHANGE
+    event: UI_SELECT.EVENT.UPDATE_MODEL
   },
   props: {
     // UI variants
+    type: {
+      type: [String, Number],
+      default: 0
+    },
     outlined: {
       type: Boolean,
       default: false
     },
     // States
     model: [String, Number],
-    selectedIndex: {
-      type: Number,
-      default: -1
-    },
     options: {
       type: Array,
       default() {
@@ -165,8 +157,10 @@ export default {
       UI_GLOBAL,
       UI_SELECT,
       $select: null,
-      selectedValue: this.model,
-      currentOption: {}
+      currentOptions: [],
+      currentOption: {},
+      currentSelectedValue: this.model,
+      currentSelectedIndex: UI_SELECT.DEFAULT_SELECTED_INDEX
     };
   },
   computed: {
@@ -182,68 +176,94 @@ export default {
         'mdc-select--no-label': this.noLabel,
         'mdc-select--with-leading-icon': this.leadingIcon || this.materialIcon
       };
+    },
+    currentSelectedLabel() {
+      let selectedLabel =
+        this.currentSelectedIndex === UI_SELECT.DEFAULT_SELECTED_INDEX
+          ? ''
+          : this.currentOption[this.optionLabel];
+
+      // TODO: Temporary solution: manual control
+      if (this.$refs.text) {
+        this.$refs.text.textContent = selectedLabel;
+      }
+
+      return selectedLabel;
+    },
+    // TODO: Temporary solution: manual control
+    isFloatAbove() {
+      return !!this.currentOption[this.optionLabel];
     }
   },
   watch: {
     model(val) {
-      this.selectedValue = val;
+      this.currentSelectedValue = val;
 
-      let selectedIndex = this.options.findIndex(
-        option => option[this.optionValue] === val
-      );
-      this.$select.selectedIndex = this.defaultLabel
-        ? selectedIndex + 1
-        : selectedIndex;
+      const index = this.setCurrentOption(this.options);
+      this.currentSelectedIndex = this.defaultLabel ? index + 1 : index;
 
+      this.$select.selectedIndex = this.currentSelectedIndex;
+
+      this.$emit(UI_SELECT.EVENT.CHANGE, this.currentSelectedValue);
       this.$emit(UI_SELECT.EVENT.SELECTED, {
-        value: val,
-        index: this.$select.selectedIndex
+        index: this.currentSelectedIndex,
+        value: this.currentOption[this.optionValue],
+        label: this.currentOption[this.optionLabel]
       });
     },
-    selectedIndex(val) {
-      this.$select.selectedIndex = val;
-
-      // this.currentOption = this.options.find((option, index) => index === val);
-      // this.selectedValue = this.currentOption[this.optionValue];
-      // this.$emit(UI_SELECT.EVENT.CHANGE, this.selectedValue);
-    },
     options(val) {
-      this.$select.selectedIndex = 0;
       this.init(val);
     }
   },
   mounted() {
     this.$select = new MDCSelect(this.$el);
+
     this.$select.listen(`MDCSelect:${UI_SELECT.EVENT.CHANGE}`, ({ detail }) => {
-      const selected = this.getSelected(detail.index);
-
-      // this.currentOption = selected;
-      // this.selectedValue = selected.value;
-
-      this.$emit(UI_SELECT.EVENT.CHANGE, selected.value);
-      this.$emit(UI_SELECT.EVENT.SELECTED, selected);
+      if (
+        this.defaultLabel
+          ? this.currentOptions.length > 1
+          : this.currentOptions.length
+      ) {
+        const selected = this.getSelected(detail.index);
+        this.$emit(UI_SELECT.EVENT.UPDATE_MODEL, selected.value);
+      }
     });
 
     this.init();
   },
   methods: {
     init(options = this.options) {
-      let currentOptions = [...options];
       // Set default option
+      let currentOptions = [...options];
       if (this.defaultLabel) {
         let defaultOption = {};
         defaultOption[this.optionLabel] = this.defaultLabel;
         defaultOption[this.optionValue] = this.defaultValue;
         currentOptions.unshift(defaultOption);
       }
+      this.currentOptions = currentOptions;
+
       // Set current option
+      const index = this.setCurrentOption();
+      this.currentSelectedIndex = index;
+    },
+    setCurrentOption(currentOptions = this.currentOptions) {
+      let index = UI_SELECT.DEFAULT_SELECTED_INDEX;
+
+      this.currentOption = {};
+      this.currentOption[this.optionLabel] = this.defaultLabel;
+      this.currentOption[this.optionValue] = this.defaultValue;
+
       for (let i = 0, len = currentOptions.length; i < len; i++) {
         let currentOption = currentOptions[i];
-        if (currentOption[this.optionValue] == this.selectedValue) {
+        if (currentOption[this.optionValue] == this.currentSelectedValue) {
+          index = i;
           this.currentOption = currentOption;
           break;
         }
       }
+
+      return index;
     },
     getSelected(index) {
       let selected = this.options[index];