+++ src/scripts/next.js
@@ -1,0 +1,32 @@
+import autoInstall from './config/auto-install';
+import multiConfigure from './config/multi-configure';
+/**
+ * Components
+ */
+import UiBottomNavigation from './components/app-bar/bottom-navigation';
+import UiBanner from './components/modal/banner';
+
+const version = require('../../package.json').version;
+
+const components = {
+  UiBottomNavigation,
+  UiBanner
+};
+
+const BalmUINext = {
+  version,
+  install(Vue, options = {}) {
+    // Configure the components' props
+    multiConfigure(components, options);
+
+    // Install the components
+    for (let key in components) {
+      let Component = components[key];
+      Vue.component(Component.name, Component);
+    }
+  }
+};
+
+autoInstall(BalmUINext);
+
+export default BalmUINext;